tidied up some code

+++ app/channels/game_channel.rb
@@ -6,9 +6,7 @@ class GameChannel < ApplicationCable::Channel
 
   def unsubscribed
     game = Game.find(params[:id])
-    if game.player_two_id == 1
-      game.destroy
-    end
+    game.destroy if game.player_two_id == 1
     stop_stream_from game
   end
 end
